+++ RELEASES.md
@@ -6,6 +6,17 @@ simplifies documenting any changes.
 # Unreleased
 
 - Send heartbeat pings to clients to detect dead connections and close them.
+  Pings are implemented at the application protocol level (i.e not WebSockets).
+- The wallet will now explicitly check for errors and responses before returning
+  when sending a request.
+- Max value ID requests will no longer respond with an IO error. However,
+  applications may misbehave as this is the reserved ID for general messages.
+
+### Breaking Changes
+
+- The network binary protocol now has a body type signifying the type of
+  message. By decoupling from RPC, we can now send generic messages that don't
+  expect a response.
 
 # Version 0.2.1 (2019-12-03)